+++ pkg/alertmanager/operator.go
@@ -21,7 +21,8 @@ import (
 	"time"
 
 	"github.com/coreos/prometheus-operator/pkg/analytics"
-	"github.com/coreos/prometheus-operator/pkg/client/monitoring/v1alpha1"
+	"github.com/coreos/prometheus-operator/pkg/client/monitoring"
+	monitoringv1 "github.com/coreos/prometheus-operator/pkg/client/monitoring/v1"
 	"github.com/coreos/prometheus-operator/pkg/k8sutil"
 	prometheusoperator "github.com/coreos/prometheus-operator/pkg/prometheus"
 
@@ -46,16 +47,14 @@ import (
 )
 
 const (
-	crdAlertmanager = v1alpha1.AlertmanagerName + "." + v1alpha1.Group
-
 	resyncPeriod = 5 * time.Minute
 )
 
 // Operator manages lify cycle of Alertmanager deployments and
 // monitoring configurations.
 type Operator struct {
-	kclient   *kubernetes.Clientset
-	mclient   *v1alpha1.MonitoringV1alpha1Client
+	kclient   kubernetes.Interface
+	mclient   monitoring.Interface
 	crdclient apiextensionsclient.Interface
 	logger    log.Logger
 
@@ -86,7 +85,7 @@ func New(c prometheusoperator.Config, logger log.Logger) (*Operator, error) {
 		return nil, errors.Wrap(err, "instantiating kubernetes client failed")
 	}
 
-	mclient, err := v1alpha1.NewForConfig(cfg)
+	mclient, err := monitoring.NewForConfig(cfg)
 	if err != nil {
 		return nil, errors.Wrap(err, "instantiating monitoring client failed")
 	}
@@ -107,13 +106,13 @@ func New(c prometheusoperator.Config, logger log.Logger) (*Operator, error) {
 
 	o.alrtInf = cache.NewSharedIndexInformer(
 		&cache.ListWatch{
-			ListFunc:  o.mclient.Alertmanagers(api.NamespaceAll).List,
-			WatchFunc: o.mclient.Alertmanagers(api.NamespaceAll).Watch,
+			ListFunc:  o.mclient.MonitoringV1().Alertmanagers(api.NamespaceAll).List,
+			WatchFunc: o.mclient.MonitoringV1().Alertmanagers(api.NamespaceAll).Watch,
 		},
-		&v1alpha1.Alertmanager{}, resyncPeriod, cache.Indexers{},
+		&monitoringv1.Alertmanager{}, resyncPeriod, cache.Indexers{},
 	)
 	o.ssetInf = cache.NewSharedIndexInformer(
-		cache.NewListWatchFromClient(o.kclient.Apps().RESTClient(), "statefulsets", api.NamespaceAll, nil),
+		cache.NewListWatchFromClient(o.kclient.AppsV1beta1().RESTClient(), "statefulsets", api.NamespaceAll, nil),
 		&v1beta1.StatefulSet{}, resyncPeriod, cache.Indexers{},
 	)
 
@@ -231,7 +230,7 @@ func (c *Operator) enqueue(obj interface{}) {
 // enqueueForNamespace enqueues all Alertmanager object keys that belong to the given namespace.
 func (c *Operator) enqueueForNamespace(ns string) {
 	cache.ListAll(c.alrtInf.GetStore(), labels.Everything(), func(obj interface{}) {
-		am := obj.(*v1alpha1.Alertmanager)
+		am := obj.(*monitoringv1.Alertmanager)
 		if am.Namespace == ns {
 			c.enqueue(am)
 		}
@@ -264,7 +263,7 @@ func (c *Operator) processNextWorkItem() bool {
 	return true
 }
 
-func (c *Operator) alertmanagerForStatefulSet(sset interface{}) *v1alpha1.Alertmanager {
+func (c *Operator) alertmanagerForStatefulSet(sset interface{}) *monitoringv1.Alertmanager {
 	key, ok := c.keyFunc(sset)
 	if !ok {
 		return nil
@@ -279,7 +278,7 @@ func (c *Operator) alertmanagerForStatefulSet(sset interface{}) *v1alpha1.Alertm
 	if !exists {
 		return nil
 	}
-	return a.(*v1alpha1.Alertmanager)
+	return a.(*monitoringv1.Alertmanager)
 }
 
 func alertmanagerNameFromStatefulSetName(name string) string {
@@ -379,7 +378,7 @@ func (c *Operator) sync(key string) error {
 		return c.destroyAlertmanager(key)
 	}
 
-	am := obj.(*v1alpha1.Alertmanager)
+	am := obj.(*monitoringv1.Alertmanager)
 	if am.Spec.Paused {
 		return nil
 	}
@@ -392,7 +391,7 @@ func (c *Operator) sync(key string) error {
 		return errors.Wrap(err, "synchronizing governing service failed")
 	}
 
-	ssetClient := c.kclient.Apps().StatefulSets(am.Namespace)
+	ssetClient := c.kclient.AppsV1beta1().StatefulSets(am.Namespace)
 	// Ensure we have a StatefulSet running Alertmanager deployed.
 	obj, exists, err = c.ssetInf.GetIndexer().GetByKey(alertmanagerKeyToStatefulSetKey(key))
 	if err != nil {
@@ -435,7 +434,7 @@ func ListOptions(name string) metav1.ListOptions {
 // create new pods.
 //
 // TODO(brancz): remove this once the 1.6 support is removed.
-func (c *Operator) syncVersion(a *v1alpha1.Alertmanager) error {
+func (c *Operator) syncVersion(a *monitoringv1.Alertmanager) error {
 	if c.config.StatefulSetUpdatesAvailable {
 		return nil
 	}
@@ -475,14 +474,14 @@ func (c *Operator) syncVersion(a *v1alpha1.Alertmanager) error {
 	return nil
 }
 
-func AlertmanagerStatus(kclient *kubernetes.Clientset, a *v1alpha1.Alertmanager) (*v1alpha1.AlertmanagerStatus, []v1.Pod, error) {
-	res := &v1alpha1.AlertmanagerStatus{Paused: a.Spec.Paused}
+func AlertmanagerStatus(kclient kubernetes.Interface, a *monitoringv1.Alertmanager) (*monitoringv1.AlertmanagerStatus, []v1.Pod, error) {
+	res := &monitoringv1.AlertmanagerStatus{Paused: a.Spec.Paused}
 
 	pods, err := kclient.Core().Pods(a.Namespace).List(ListOptions(a.Name))
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "retrieving pods of failed")
 	}
-	sset, err := kclient.Apps().StatefulSets(a.Namespace).Get(statefulSetNameFromAlertmanagerName(a.Name), metav1.GetOptions{})
+	sset, err := kclient.AppsV1beta1().StatefulSets(a.Namespace).Get(statefulSetNameFromAlertmanagerName(a.Name), metav1.GetOptions{})
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "retrieving stateful set failed")
 	}
@@ -539,7 +538,7 @@ func (c *Operator) destroyAlertmanager(key string) error {
 	*sset.Spec.Replicas = 0
 
 	// Update the replica count to 0 and wait for all pods to be deleted.
-	ssetClient := c.kclient.Apps().StatefulSets(sset.Namespace)
+	ssetClient := c.kclient.AppsV1beta1().StatefulSets(sset.Namespace)
 
 	if _, err := ssetClient.Update(sset); err != nil {
 		return errors.Wrap(err, "updating statefulset for scale-down failed")
@@ -570,20 +569,7 @@ func (c *Operator) destroyAlertmanager(key string) error {
 
 func (c *Operator) createCRDs() error {
 	crds := []*extensionsobj.CustomResourceDefinition{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: crdAlertmanager,
-			},
-			Spec: extensionsobj.CustomResourceDefinitionSpec{
-				Group:   v1alpha1.Group,
-				Version: v1alpha1.Version,
-				Scope:   extensionsobj.NamespaceScoped,
-				Names: extensionsobj.CustomResourceDefinitionNames{
-					Plural: v1alpha1.AlertmanagerName,
-					Kind:   v1alpha1.AlertmanagersKind,
-				},
-			},
-		},
+		k8sutil.NewAlertmanagerCustomResourceDefinition(),
 	}
 
 	crdClient := c.crdclient.ApiextensionsV1beta1().CustomResourceDefinitions()
@@ -596,20 +582,12 @@ func (c *Operator) createCRDs() error {
 	}
 
 	// We have to wait for the CRDs to be ready. Otherwise the initial watch may fail.
-	return k8sutil.WaitForCRDReady(c.kclient.CoreV1().RESTClient(), v1alpha1.Group, v1alpha1.Version, v1alpha1.AlertmanagerName)
+	return k8sutil.WaitForCRDReady(c.mclient.MonitoringV1().Alertmanagers(api.NamespaceAll).List)
 }
 
 func (c *Operator) createTPRs() error {
 	tprs := []*extensionsobjold.ThirdPartyResource{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "alertmanager." + v1alpha1.Group,
-			},
-			Versions: []extensionsobjold.APIVersion{
-				{Name: v1alpha1.Version},
-			},
-			Description: "Managed Alertmanager cluster",
-		},
+		k8sutil.NewAlertmanagerTPRDefinition(),
 	}
 	tprClient := c.kclient.Extensions().ThirdPartyResources()
 
@@ -621,5 +599,5 @@ func (c *Operator) createTPRs() error {
 	}
 
 	// We have to wait for the TPRs to be ready. Otherwise the initial watch may fail.
-	return k8sutil.WaitForCRDReady(c.kclient.CoreV1().RESTClient(), v1alpha1.Group, v1alpha1.Version, v1alpha1.AlertmanagerName)
+	return k8sutil.WaitForCRDReady(c.mclient.MonitoringV1alpha1().Alertmanagers(api.NamespaceAll).List)
 }