+++ test/framework/framework.go
@@ -15,6 +15,9 @@
 package framework
 
 import (
+	"bytes"
+	"fmt"
+	"io"
 	"net/http"
 	"testing"
 	"time"
@@ -27,19 +30,20 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/coreos/prometheus-operator/pkg/client/monitoring/v1alpha1"
+	monitoringv1 "github.com/coreos/prometheus-operator/pkg/client/monitoring/v1"
 	"github.com/coreos/prometheus-operator/pkg/k8sutil"
 	"github.com/pkg/errors"
 )
 
 type Framework struct {
 	KubeClient     kubernetes.Interface
-	MonClient      *v1alpha1.MonitoringV1alpha1Client
+	MonClient      monitoringv1.MonitoringV1Interface
 	HTTPClient     *http.Client
 	MasterHost     string
 	Namespace      *v1.Namespace
 	OperatorPod    *v1.Pod
 	DefaultTimeout time.Duration
+	operatorLogs   *bytes.Buffer
 }
 
 // Setup setups a test framework and returns it.
@@ -59,7 +63,7 @@ func New(ns, kubeconfig, opImage string) (*Framework, error) {
 		return nil, errors.Wrap(err, "creating http-client failed")
 	}
 
-	mclient, err := v1alpha1.NewForConfig(config)
+	mclient, err := monitoringv1.NewForConfig(config)
 	if err != nil {
 		return nil, errors.Wrap(err, "creating monitoring client failed")
 	}
@@ -76,6 +80,7 @@ func New(ns, kubeconfig, opImage string) (*Framework, error) {
 		HTTPClient:     httpc,
 		Namespace:      namespace,
 		DefaultTimeout: time.Minute,
+		operatorLogs:   bytes.NewBuffer(nil),
 	}
 
 	err = f.Setup(opImage)
@@ -90,6 +95,49 @@ func (f *Framework) Setup(opImage string) error {
 	if err := f.setupPrometheusOperator(opImage); err != nil {
 		return errors.Wrap(err, "setup prometheus operator failed")
 	}
+
+	go f.retryOperatorLogs()
+
+	return nil
+}
+
+func (f *Framework) retryOperatorLogs() {
+	for {
+		err := f.recordOperatorLogs()
+		if err != nil {
+			errtxt := fmt.Sprintf("\n--- There was an error capturing logs (%s), retrying... ---\n", err)
+			f.operatorLogs.Write([]byte(errtxt))
+			time.Sleep(time.Second)
+		}
+	}
+}
+
+func (f *Framework) recordOperatorLogs() error {
+	deploy, err := f.KubeClient.AppsV1beta1().Deployments(f.Namespace.Name).Get("prometheus-operator", metav1.GetOptions{})
+	if err != nil {
+		return err
+	}
+	opts := metav1.ListOptions{LabelSelector: fields.SelectorFromSet(fields.Set(deploy.Spec.Template.ObjectMeta.Labels)).String()}
+	list, err := f.KubeClient.CoreV1().Pods(f.Namespace.Name).List(opts)
+	if err != nil {
+		return err
+	}
+
+	if len(list.Items) != 1 {
+		return fmt.Errorf("1 Prometheus Operator Pod expected, but found %d", len(list.Items))
+	}
+
+	r, err := f.KubeClient.CoreV1().Pods(f.Namespace.Name).GetLogs(list.Items[0].Name, &v1.PodLogOptions{Follow: true}).Stream()
+	if err != nil {
+		return err
+	}
+	defer r.Close()
+
+	_, err = io.Copy(f.operatorLogs, r)
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
 
@@ -137,17 +185,7 @@ func (f *Framework) setupPrometheusOperator(opImage string) error {
 	}
 	f.OperatorPod = &pl.Items[0]
 
-	err = k8sutil.WaitForCRDReady(f.KubeClient.Core().RESTClient(), v1alpha1.Group, v1alpha1.Version, v1alpha1.PrometheusName)
-	if err != nil {
-		return err
-	}
-
-	err = k8sutil.WaitForCRDReady(f.KubeClient.Core().RESTClient(), v1alpha1.Group, v1alpha1.Version, v1alpha1.ServiceMonitorName)
-	if err != nil {
-		return err
-	}
-
-	return k8sutil.WaitForCRDReady(f.KubeClient.Core().RESTClient(), v1alpha1.Group, v1alpha1.Version, v1alpha1.AlertmanagerName)
+	return nil
 }
 
 func (ctx *TestCtx) SetupPrometheusRBAC(t *testing.T, ns string, kubeClient kubernetes.Interface) {
@@ -166,6 +204,8 @@ func (ctx *TestCtx) SetupPrometheusRBAC(t *testing.T, ns string, kubeClient kube
 
 // Teardown tears down a previously initialized test environment.
 func (f *Framework) Teardown() error {
+	fmt.Println("Prometheus Operator Logs Captured: \n\n", f.operatorLogs.String())
+
 	if err := f.KubeClient.Core().Services(f.Namespace.Name).Delete("prometheus-operated", nil); err != nil && !k8sutil.IsResourceNotFoundError(err) {
 		return err
 	}